Move logic into client_request for consistency
